@@ -1,6 +1,6 @@
 {
   "name": "ng-deploy-oss",
-  "version": "18.0.0",
+  "version": "19.0.0",
   "description": "Deploy Angular apps to aliyun OSS, qiniu, upyun using the Angular CLI.",
   "keywords": [
     "schematics",
@@ -33,39 +33,41 @@
     "test": "jest --watch",
     "test:ci": "jest",
     "lint": "eslint .",
-    "build": "tsc scripts/build.ts; node ./scripts/build.js",
+    "build": "tsx scripts/build.ts",
     "build:test": "npm run build -- --test",
     "release": "npm run build -- --release"
   },
   "dependencies": {
-    "@angular-devkit/architect": "^0.1800.0",
-    "@angular-devkit/build-angular": "18.0.0",
-    "@angular-devkit/core": "^18.0.0",
-    "@angular-devkit/schematics": "^18.0.0",
-    "ali-oss": "^6.20.0",
+    "@angular-devkit/architect": "^0.1900.1",
+    "@angular-devkit/build-angular": "19.0.1",
+    "@angular-devkit/core": "^19.0.1",
+    "@angular-devkit/schematics": "^19.0.1",
+    "ali-oss": "^6.21.0",
     "chalk": "^5.3.0",
     "fs-extra": "^11.2.0",
-    "inquirer": "^9.2.23",
+    "inquirer": "^12.1.0",
     "inquirer-autocomplete-prompt": "^3.0.1",
-    "qiniu": "^7.12.0",
-    "typescript": "~5.0.4",
+    "qiniu": "^7.14.0",
+    "typescript": "~5.6.2",
     "upyun": "^3.4.6"
   },
   "devDependencies": {
-    "@schematics/angular": "^18.0.0",
-    "@types/ali-oss": "^6.16.8",
+    "@schematics/angular": "^19.0.1",
+    "@types/ali-oss": "^6.16.11",
     "@types/fs-extra": "^11.0.4",
     "@types/inquirer": "9.0.7",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^18.19.34",
-    "@types/rimraf": "^3.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "@typescript-eslint/parser": "^6.10.0",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^20.17.7",
+    "@types/rimraf": "^4.0.5",
+    "@types/upyun": "^3.4.0",
+    "@typescript-eslint/eslint-plugin": "^8.15.0",
+    "@typescript-eslint/parser": "^8.15.0",
     "eslint": "^8.53.0",
     "jest": "^29.7.0",
-    "prettier": "^2.8.8",
-    "rimraf": "^3.0.2",
-    "ts-jest": "^29.1.4"
+    "prettier": "^3.3.3",
+    "rimraf": "^6.0.1",
+    "ts-jest": "^29.2.5",
+    "tsx": "^4.19.2"
   },
   "jest": {
     "roots": [
@@ -74,5 +76,6 @@
     "transform": {
       "^.+\\.tsx?$": "ts-jest"
     }
-  }
+  },
+  "packageManager": "yarn@4.5.2"
 }
